requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlmodel==0.0.14
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==46.0.1

# AWS/S3 Storage
boto3==1.34.34
botocore==1.34.34

# Background Tasks
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.0

# Data Validation
pydantic==2.5.2
email-validator==2.1.0.post1
pydantic-settings==2.10.1

# Environment Variables
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-fastapi-instrumentator==6.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
coverage==7.3.2

# AI/ML Processing (placeholders for future LoRA models)
Pillow==10.1.0
numpy==1.25.2
python-magic==0.4.27

# Monitoring & Logging
structlog==23.2.0